develop/pyproject.toml
# Copyright 2024 ECSDevs
# 
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
# 
#     https://www.apache.org/licenses/LICENSE-2.0
# 
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "syncfile"
version = "0.3.1.3"
authors = [
  { name="ECSDevs", email="psbash@groupmail.com" },
]
description = "Simple file synchronization service (clients in the same package)"
readme = "README.md"
requires-python = ">=3.8"
license = {file = "LICENCE"}
maintainers = [
  {name = "ECSDevs", email = "2438926613@qq.com" }
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: Apache Software License",
  "Operating System :: OS Independent",
]
dependencies = [
  "dnspython",
  "httpx",
  "pydantic"
]

[project.urls]
"Homepage" = "https://github.com/ECSDevs/FileSync"
"Bug Tracker" = "https://github.com/ECSDevs/FileSync/issues"
"Support Me" = "https://afdian.net/a/Syncfile_transfer_tool"
